requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-texttospeech==2.21.1
python-multipart==0.0.20
